fix: Resolve CI test failures
- Fix verify_consistency missing RETURN statement in conflict resolution
- Fix round() type casting for PostgreSQL compatibility
- Fix history table partitioning INSERT column mismatch
- Remove indexes on non-existent columns in commits table
- Remove foreign key constraints on non-existent commit_id column
- Add test-with-fixes workflow for step-by-step debugging

These changes ensure tests can run in clean CI environments without
assuming specific table structures exist.

+++ sql/052_performance_monitoring.sql
@@ -447,9 +447,9 @@ SELECT
     ROUND(MIN(duration_ms), 2) as min_duration_ms,
     ROUND(MAX(duration_ms), 2) as max_duration_ms,
     ROUND(STDDEV(duration_ms), 2) as stddev_duration_ms,
-    ROUND(percentile_cont(0.5) WITHIN GROUP (ORDER BY duration_ms), 2) as median_duration_ms,
-    ROUND(percentile_cont(0.95) WITHIN GROUP (ORDER BY duration_ms), 2) as p95_duration_ms,
-    ROUND(percentile_cont(0.99) WITHIN GROUP (ORDER BY duration_ms), 2) as p99_duration_ms
+    ROUND(percentile_cont(0.5) WITHIN GROUP (ORDER BY duration_ms)::numeric, 2) as median_duration_ms,
+    ROUND(percentile_cont(0.95) WITHIN GROUP (ORDER BY duration_ms)::numeric, 2) as p95_duration_ms,
+    ROUND(percentile_cont(0.99) WITHIN GROUP (ORDER BY duration_ms)::numeric, 2) as p99_duration_ms
 FROM pggit.performance_metrics
 WHERE started_at >= now() - interval '24 hours'
 GROUP BY operation_type;
